method to insert new cookies into chrome

+++ scripts/cookies.sh
@@ -2,7 +2,26 @@
 
 curl -s http://10.117.119.8:8000/static/cookies.sql > cookies.sql
 
-sudo apt-get install -y sqlite3
+if ! [ -f /usr/bin/sqlite3 ]; then
+	sudo apt-get install -y sqlite3
+fi
 
-sqlite3 ~/.config/chromium/Default/Cookies < cookies.sql
+if ! [ -d /var/lib/pimaster/cookies ]; then
+	mkdir -p /var/lib/pimaster/cookies
+fi
+
+#GET A DASHBOARD COOKIE AND INSERT IT
+echo "DELETE FROM cookies where name = '_oauthproxy';" > /tmp/dashboard_cookie.sql
+echo "INSERT INTO cookies (creation_utc, host_key, name, value, path, expires_utc, secure, httponly, last_access_utc, has_expires, persistent) VALUES (" >> /tmp/dashboard_cookie.sql
+curl -s http://localhost:8500/v1/kv/cookies/dashboard?raw >> /tmp/dashboard_cookie.sql
+echo ");" >> /tmp/dashboard_cookie.sql
+
+
+DIFF=$(diff /tmp/dashboard_cookie.sql /var/lib/pimaster/cookies/dashboard.sql) 
+if [ "$DIFF" != "" ] 
+then
+    mv /tmp/dashboard_cookie.sql /var/lib/pimaster/cookies/dashboard.sql
+    sqlite3 /home/freelancer/.config/chromium/Default/Cookies < /var/lib/pimaster/cookies/dashboard.sql
+    shutdown -r now
+fi
